backend/requirements.txt
Flask==2.0.1
spotipy==2.19.0
python-dotenv==0.19.0
matplotlib==3.4.2
pandas==1.3.2
